% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation Plot}
\usage{
correlation()
}
\description{
This function analyses the correlation between the numeric variables in the dataset and returns the correlation plot
based on the correlation matrix.
Note: this function requires the package corrplot.
}
\details{
This function needs the argument air_safety
}
