% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFInc.R
\name{nFInc}
\alias{nFInc}
\title{This function requires as the input the number associated to one airline and returns the number of fatal incidents
experienced by that airline between 2000 and 2014.}
\usage{
nFInc(airline)
}
\description{
This function requires as the input the number associated to one airline and returns the number of fatal incidents
experienced by that airline between 2000 and 2014.
}
