% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\name{reg}
\alias{reg}
\title{Linear Regressions}
\usage{
reg()
}
\description{
This function runs two linear regression models: the first one is a simple linear regression of the number of
fatal accidents between 2000 and 2014 on the number of incidents between 2000 and 2014; the second one is a
multiple linear regression of the number of fatal accidents between 2000 and 2014 on the number of incidents
between 2000 and 2014 and the number of fatalities between 2000 and 2014.
As a result, it returns the summary of the two models.
This function needs the argument air_safety.
}
