% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFInc.R
\name{nFInc}
\alias{nFInc}
\title{Number of fatal accidents}
\usage{
nFInc(airline)
}
\arguments{
\item{airline}{integer}
}
\description{
This function requires as the input the number associated to one airline and returns the number of fatal accidents
experienced by that airline between 2000 and 2014.
}
\details{
This function needs the argument airline.
}
