% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1.R
\name{plot1}
\alias{plot1}
\title{Barplot}
\usage{
plot1(x, y, tipe = "bar", air_safety)
}
\arguments{
\item{x}{factor}

\item{y}{integer}

\item{tipe}{character}

\item{air_safety}{data.frame}
}
\value{
piechart
}
\description{
This function creates a barplot representing the number of fatal accidents for each airline in the two periods:
1985-1999 and 2000-2014.
Note: this function requires the package plotly.
}
\details{
This function needs the argument air_safety
}
