% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\name{reg}
\alias{reg}
\title{This function runs two linear regression models:
- the first one is a simple linear regression of fatal_accidents_00_14 on incidents_00_14
- the second one is a multiple linear regression of fatal_accidents_00_14 on incidents_00_14 and fatalities_00_14.
As a result, it returns the summary of the two models.}
\usage{
reg(air_safety)
}
\arguments{
\item{air_safety}{data.frame}
}
\description{
This function runs two linear regression models:
- the first one is a simple linear regression of fatal_accidents_00_14 on incidents_00_14
- the second one is a multiple linear regression of fatal_accidents_00_14 on incidents_00_14 and fatalities_00_14.
As a result, it returns the summary of the two models.
}
